requirements.txt
aiofiles==24.1.0
aiogram==3.14.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
asttokens==2.4.1
asyncpg==0.30.0
attrs==24.2.0
Babel==2.13.1
backcall==0.2.0
black==23.11.0
bracex==2.4
build==1.2.2.post1
CacheControl==0.14.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
codecov==2.1.13
colorama==0.4.6
coverage==7.3.2
crashtest==0.4.1
cryptography==43.0.3
decorator==5.1.1
deprecation==2.1.0
distlib==0.3.9
dulwich==0.21.7
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.16.1
freezegun==1.3.0
frozenlist==1.5.0
ghp-import==2.1.0
griffe==0.38.0
grpcio==1.59.3
grpcio-tools==1.59.3
idna==3.6
iniconfig==2.0.0
installer==0.7.0
ipython==8.12.3
isort==5.12.0
jaraco.classes==3.4.0
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.2
keyring==24.3.1
magic-filter==1.0.12
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mergedeep==1.3.4
mkdocs==1.5.3
mkdocs-autorefs==0.5.0
mkdocs-include-markdown-plugin==6.0.4
mkdocs-material==9.4.14
mkdocs-material-extensions==1.3.1
mkdocstrings==0.24.0
mkdocstrings-python==1.7.5
more-itertools==10.5.0
msgpack==1.1.0
multidict==6.1.0
mypy==1.7.1
mypy-extensions==1.0.0
mypy-protobuf==3.5.0
packaging==23.2
paginate==0.5.6
parso==0.8.3
pathspec==0.11.2
pexpect==4.9.0
pickleshare==0.7.5
pkginfo==1.11.2
platformdirs==4.0.0
pluggy==1.3.0
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
prompt-toolkit==3.0.41
propcache==0.2.0
protobuf==4.25.1
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.17.2
pymdown-extensions==10.5
pyproject_hooks==1.2.0
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
pytest-deadfixtures==2.2.1
pytest-freezegun==0.4.2
pytest-mock==3.12.0
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
pyyaml_env_tag==0.1
RapidFuzz==3.10.1
regex==2023.10.3
requests==2.31.0
requests-toolbelt==1.0.0
ruff==0.1.6
SecretStorage==3.3.3
setuptools==69.0.2
shellingham==1.5.4
six==1.16.0
stack-data==0.6.3
termynal==0.11.1
tinkoff==0.1.1
tokenize-rt==5.2.0
tomlkit==0.12.3
traitlets==5.14.0
trove-classifiers==2024.10.21.16
types-cachetools==5.3.0.7
types-protobuf==4.24.0.4
types-python-dateutil==2.8.19.14
types-PyYAML==6.0.12.12
types-requests==2.31.0.10
typing_extensions==4.8.0
urllib3==2.1.0
virtualenv==20.27.1
watchdog==3.0.0
wcmatch==8.5
wcwidth==0.2.12
yarl==1.17.1
